Must have: Average age of infected people
A general calculation of the average age of infected people was carried out, but also through certain fields or necessary variables such as: age, sex, infected or not infected, returned home or hospitalized, level of medical unit, diabetes, asthma and obesity. Then, conclusions were made from the section according to the calculations analyzed.